feat: Add ts-morph for query resolution of fragments
This introduces using ts-morph to parse and resolve extracted queries
from defineQuery() calls and groq`` template literals.
This allows for resolving queries which uses string interpolation
to embed fragments in the query.

Example:

  const subQuery = defineQuery(`{ name }`)

  const query = defineQuery(`{
    _id,
    author->${subQuery}
  }`)

will resolve to:

  {
    _id,
    author->{ name }
  }

Since we leverage ts-morph to parse the extracted code, this will also
work if `subQuery` is defined in a separate file.

To use ts-morph we need to create a Project - this is stored in a cache
based on the ts-config file to ensure that we don't need to do all the
parsing on each iteration. The extractAllDefineQuery() will get the
source and path of the current file, and ensure that it is always up to
date in the Project.

We implement some logic to resolve the nearest ts config file from the
current file. This should hopefully ensure everything resolves correctly
in a mono repo.

I had to update Prettier to get it working on my machine.