feat: Add copy query command
Small QOL improvement which adds a 'Copy Query' command along side
'Execute Query'. Makes it easy to copy the query if you want to execute
it in Vision or send it to someone.

Looks like a few of the files were not formatted according to the
Prettier rules, which caused a slightly larger change set than desired.